+++ server/routers/__init__.py
@@ -2,3 +2,4 @@ from .health import router as health_router
 from .scene import router as scene_router
 from .proxy import router as proxy_router
 from .llm_callback import router as llm_callback_router
+from .llm_debug import router as llm_debug_router