+++ server/services/__init__.py
@@ -1,3 +1,5 @@
 from .scene_service import build_scene_list
 from .proxy_service import proxy_request
 from .llm_channel import chat_completion_stream
+from .llm_service import llm_service
+from .rag_service import rag_service